--- a/src/modules/logger/logger.h
+++ b/src/modules/logger/logger.h
@@ -79,27 +79,17 @@ inline bool operator&(SDLogProfileMask a, SDLogProfileMask b)
 	return static_cast<int32_t>(a) & static_cast<int32_t>(b);
 }
 
-struct LoggerSubscription {
-	int fd[ORB_MULTI_MAX_INSTANCES]; ///< uorb subscription. The first fd is also used to store the interval if
-	/// not subscribed yet (-interval - 1)
-	const orb_metadata *metadata = nullptr;
-	uint8_t msg_ids[ORB_MULTI_MAX_INSTANCES];
+static constexpr uint8_t MSG_ID_INVALID = UINT8_MAX;
 
-	LoggerSubscription() {}
+struct LoggerSubscription : public uORB::SubscriptionInterval {
 
-	LoggerSubscription(int fd_, const orb_metadata *metadata_) :
-		metadata(metadata_)
-	{
-		fd[0] = fd_;
+	uint8_t msg_ids{MSG_ID_INVALID};
 
-		for (int i = 1; i < ORB_MULTI_MAX_INSTANCES; i++) {
-			fd[i] = -1;
-		}
+	LoggerSubscription() = default;
 
-		for (int i = 0; i < ORB_MULTI_MAX_INSTANCES; i++) {
-			msg_ids[i] = (uint8_t) - 1;
-		}
-	}
+	LoggerSubscription(const orb_metadata *meta, unsigned interval = 0, uint8_t instance = 0) :
+		uORB::SubscriptionInterval(meta, interval, instance)
+	{}
 };
 
 class Logger : public ModuleBase<Logger>
@@ -142,7 +132,8 @@ public:
 	 * @param interval limit rate if >0, otherwise log as fast as the topic is updated.
 	 * @return true on success
 	 */
-	bool add_topic(const char *name, unsigned interval = 0);
+	bool add_topic(const char *name, unsigned interval = 0, uint8_t instance = 0);
+	bool add_topic_multi(const char *name, unsigned interval = 0);
 
 	/**
 	 * add a logged topic (called by add_topic() above).
@@ -150,7 +141,7 @@ public:
 	 * and sets the file descriptor of LoggerSubscription accordingly
 	 * @return the newly added subscription on success, nullptr otherwise
 	 */
-	LoggerSubscription *add_topic(const orb_metadata *topic);
+	LoggerSubscription *add_topic(const orb_metadata *topic, unsigned interval = 0, uint8_t instance = 0);
 
 	/**
 	 * request the logger thread to stop (this method does not block).
@@ -207,7 +198,7 @@ private:
 	 * Write an ADD_LOGGED_MSG to the log for a given subscription and instance.
 	 * _writer.lock() must be held when calling this.
 	 */
-	void write_add_logged_msg(LogType type, LoggerSubscription &subscription, int instance);
+	void write_add_logged_msg(LogType type, LoggerSubscription &subscription);
 
 	/**
 	 * Create logging directory
@@ -285,13 +276,7 @@ private:
 
 	void write_changed_parameters(LogType type);
 
-	inline bool copy_if_updated_multi(int sub_idx, int multi_instance, void *buffer, bool try_to_subscribe);
-
-	/**
-	 * Check if a topic instance exists and subscribe to it
-	 * @return true when topic exists and subscription successful
-	 */
-	bool try_to_subscribe_topic(LoggerSubscription &sub, int multi_instance);
+	bool copy_if_updated(int sub_idx, void *buffer, bool try_to_subscribe);
 
 	/**
 	 * Write exactly one ulog message to the logger and handle dropouts.