--- a/src/modules/mavlink/mavlink_mission.h
+++ b/src/modules/mavlink/mavlink_mission.h
@@ -46,7 +46,8 @@
 #pragma once
 
 #include <dataman/dataman.h>
-#include <uORB/uORB.h>
+#include <uORB/Subscription.hpp>
+#include <uORB/topics/mission_result.h>
 
 #include "mavlink_bridge_header.h"
 #include "mavlink_rate_limiter.h"
@@ -124,8 +125,7 @@ private:
 
 	static bool		_transfer_in_progress;			///< Global variable checking for current transmission
 
-	int			_offboard_mission_sub{-1};
-	int			_mission_result_sub{-1};
+	uORB::SubscriptionBase	_mission_result_sub{ORB_ID(mission_result)};
 
 	orb_advert_t		_offboard_mission_pub{nullptr};