--- a/src/modules/mavlink/mavlink_orb_subscription.h
+++ b/src/modules/mavlink/mavlink_orb_subscription.h
@@ -93,7 +93,7 @@ public:
 
 private:
 
-	uORB::SubscriptionBase	_sub;
+	uORB::Subscription	_sub;
 };