@dagar 2019-01-09
Note - this is still early and needs quite a bit more work. I'd mainly like to start getting feedback on the overall design/direction and any other thoughts.