@julianoes 2019-01-09
Nice work! I think this is viable but I might not have all the context because I don't know why these file descriptors were used in the first place. I looked at the diff but it's quite big and seems to contain a lot of overall cleanup, so it's hard to see the meat of the change.

What I wonder is if this changes anything about the semantics of polling on topics. I was always assuming that was one of the reasons file descriptors were used.