@TSC21 2019-01-09
What's the impact of this on the FastRTPS communications?