@dagar 2019-01-09
> Nice work! I think this is viable but I might not have all the context because I don't know why these file descriptors were used in the first place. 

I don't know the full history, but in theory it's nice if you can rely on existing system facilities for nearly everything (open/close/read/write/poll). We've gradually grown out various levels of wrapper and obfuscation though.

> I looked at the diff but it's quite big and seems to contain a lot of overall cleanup, so it's hard to see the meat of the change.

The first commit is the uORB change itself, the rest are updates to use it. In particular look at uORBDeviceNode.

> 
> What I wonder is if this changes anything about the semantics of polling on topics. I was always assuming that was one of the reasons file descriptors were used.

For now we'd continue using a regular subscription and poll, but we'll be able to do it through these eventually. It could be something like an alternate update call that blocks (eg `updateBlocking()`.