@dagar 2019-01-09
> What's the impact of this on the FastRTPS communications?

None initially, but uORB::Subscription could grow a blocking update method that would then be good to consider within the RTPS bridge.