@dagar 2019-01-09
> - I'd make the default `Subscription` class not contain the topic data, and add a `SubscriptionData` (or whatever name) class that contains it, to encourage the use of subscription w/o data (for the cases where we can keep the data on the stack for example).

Good idea.

> - can we avoid the virtual overhead of Subscription?

Probably not in the first pass, but I think we'll be able to get there eventually. At the moment SubscriptionBase is fairly flexible to support the mavlink and logger usage, but a template version intended for embedding directly in modules (90% of the system orb usage) can be much simpler.

> - Detail: if we could keep the `add_topic` in Logger generic to add all instances automatically, I'd prefer that. No big deal though.

Kind of a side note, but it seems to me that topics should be explicitly multi-capable or single only. I think I'll come back to this idea later. For logger I was trying to avoid creating 4x SubscriptionBases for every topic when only a small portion of them are multi. Any suggestions?

> >  I also have a template variation in mind on the actual ORB_ID type that would remove for need for storing metadata in memory for a total of 8 bytes per subscription.
> 
> I was also thinking of doing that but never got to it so far.

We need to update the generation to produce something like the enum class you created for params. I was also hoping to find a way to tie the topic and data structure together (not sure if this will be controversial) to make the declaration trivial. For example `uORB::Subscription<px4::vehicle_status>` would use `vehicle_status_s` references for update, and perhaps even have a getter that simply returns a vehicle_status_s by value. 

> 
> > What I wonder is if this changes anything about the semantics of polling on topics.
> 
> I assume we'd still use the existing interface for that?

Yes for now, but I'd like to add a blocking update function here as well (later). I think we can do slightly better than the performance of NuttX poll, while making it hard to misuse.