@dagar 2019-06-07
Implemented! https://github.com/PX4/Firmware/pull/12040